small analysis of monkey responses to frozen RDM noise
The program fetches the data from neuralsignal.org, extracts the choices of the
monkey and plots the fraction that one response was given for experimental
blocks in which the random seed was constant (frozen noise), or not.

There is a clear difference in the responses: In response to frozen noise the
monkey has more stereotypical choices, as one would expect. Note that it is
unclear from the data description whether the noise was frozen also across
experimental blocks (corresponding to single neurons). Looking at the data,
however, it appears extremely unlikely that always the same RDM stimulus was
used, because in some blocks the monkey chose almost exclusivel response 1
whereas in others it chose almost exclusively response 2. Given that there were
always trials with varying random seed and non-0% coherence interleaved it
appears highly implausible that the monkey would switch its response behaviour
so drastically without relation to the stimulus.

I thought about trying to statistically infer whether responses to frozen noise
are indeed more sterotypical. I have identified PyStan as an option to implement
this. You can model the data hierarchically with a Binomial distribution of the
responses on the level of experiment blocks and a Beta distribution of
underlying response probabilities per block which should differ across the two
noise conditions (frozen and variable). It should be easy to implement such a
model, but Stan does not allow model comparison, for example, for comparing a
model which allows for different Beta-parameters for the two conditions vs. a
model which takes common Beta-parameters. I could only fit the model separately
to the two conditions and compare the resulting posteriors of the
Beta-parameters. Given the clear differences in the responses this should give a
clear difference in the posteriors, but interpretation would still be hard,
because then you have to argue based on the Beta-parameter values, i.e., the
shape of the Beta distribution which perhaps is unintuitive for many people.